update to GUi based on tutorial
